+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 
 ## 0.4.1.0 (UNRELEASED)
 
-- `Rib.Pandoc`: Export `render'` and `renderInlines'` (the non-Lucid versions)
-- `Rib.Pandoc`: Re-export `Text.Pandoc.Readers` so the library user does not
-  have to directly depend on `pandoc` only to render its documents.
+- `Rib.Pandoc`
+  - Export `render'` and `renderInlines'` (the non-Lucid versions)
+  - Re-export `Text.Pandoc.Readers` so the library user does not have to directly depend on `pandoc` only to render its documents.
+  - Add `getToC` returning table of contents of a document
+
 - `Rib.App`: The `run` funtion now takes two more arguments, specifying the input and output directory, which are no longer hardcoded.
 - `Rib.Simple`: add LaTeX to default list of readers
 - `Rib.Server`: Remove ".html" detection magic from URLs
+- Upgraded to pandoc 2.8
 
 ## 0.3.0.0